fix syntax errors

+++ src/hikogui/widgets/toggle_widget.hpp
@@ -193,7 +193,7 @@ public:
             hilet pip_size = theme<prefix / "pip">.size(this);
             hi_axiom(pip_size.width() == pip_size.height());
 
-            _pip_circle = align(button_square, circle{pip_size.width() / 2}, alignment::middle_center());
+            _pip_circle = align(button_square, circle{narrow_cast<float>(pip_size.width() / 2)}, alignment::middle_center());
 
             hilet pip_to_button_margin_x2 = _button_rectangle.height() - narrow_cast<int>(_pip_circle.diameter());
             _pip_move_range = _button_rectangle.width() - narrow_cast<int>(_pip_circle.diameter()) - pip_to_button_margin_x2;
